refactor: fix layer violations and extract TokenService from AuthService (#46)
- Fix AuthService directly accessing UserRepository (audit #7): add
  setEmailVerificationToken, verifyEmail, findFullById to UserService
- Fix ProfileService directly accessing UserRepository (audit #8): add
  findProfileById, updateProfile, findWithPasswordHash, updatePasswordHash,
  softDelete to UserService; remove CacheService dep from ProfileService
- Fix excessive constructor params in AuthService (audit #33): extract
  TokenService handling JWT generation, refresh token CRUD, blacklisting,
  and logout; AuthService reduced from 8 to 4 dependencies
- Update OnboardingService and ScheduledTasksService to use service layer
  instead of direct repository access
- Stop exporting UserRepository from UserModule
